+++ sqlite3/lib/sqlite3.dart
@@ -2,7 +2,6 @@
 library sqlite3;
 
 // Hide common base classes that have more specific ffi-APIs.
-export 'common.dart'
-    hide CommonPreparedStatement, CommonDatabase, CommmonSqlite3;
+export 'common.dart' hide CommonPreparedStatement, CommonDatabase;
 
 export 'src/ffi/api.dart';