+++ sqlite3/lib/src/result_set.dart
@@ -1,14 +1,12 @@
 import 'dart:collection';
 
 import 'package:collection/collection.dart';
-import 'package:meta/meta.dart';
 
 /// Base class for result sets.
 ///
 /// Result sets are either completely materialized ([ResultSet] with all rows
 /// being directly available), or executed row-by-row ([IteratingCursor]).
-@sealed
-abstract class Cursor {
+sealed class Cursor {
   /// The column names of this query, as returned by `sqlite3`.
   final List<String> columnNames;
 
@@ -39,7 +37,7 @@ abstract class IteratingCursor extends Cursor implements Iterator<Row> {
 }
 
 /// Stores the full result of a select statement.
-class ResultSet extends Cursor
+final class ResultSet extends Cursor
     with
         ListMixin<Row>,
         NonGrowableListMixin<Row> // ignore: prefer_mixin
@@ -72,7 +70,7 @@ class ResultSet extends Cursor
 /// value of a column by its name.
 /// The [columnAt] method may be used to obtain the value of a column by its
 /// index.
-class Row
+final class Row
     with
         // ignore: prefer_mixin
         UnmodifiableMapMixin<String, dynamic>,
@@ -135,7 +133,7 @@ class Row
   }
 }
 
-class _ResultIterator implements Iterator<Row> {
+final class _ResultIterator implements Iterator<Row> {
   final ResultSet result;
   int index = -1;