+++ sqlite3/lib/src/sqlite3.dart
@@ -1,9 +1,7 @@
 import 'database.dart';
 
-typedef CommmonSqlite3 = CommonSqlite3;
-
 /// Provides access to `sqlite3` functions, such as opening new databases.
-abstract class CommonSqlite3 {
+abstract interface class CommonSqlite3 {
   /// The version of the sqlite3 library in used.
   Version get version;
 
@@ -39,7 +37,7 @@ abstract class CommonSqlite3 {
 }
 
 /// Version information about the sqlite3 library in use.
-class Version {
+final class Version {
   /// A textual description of this sqlite version, such as '3.32.2'.
   final String libVersion;