+++ sqlite3/test/common/prepared_statement.dart
@@ -227,6 +227,24 @@ void testPreparedStatements(
     opened.dispose();
   });
 
+  test('handles recompilations', () {
+    final opened = sqlite3.openInMemory()
+      ..execute('create table t (c1)')
+      ..execute('insert into t values (1)');
+    addTearDown(opened.dispose);
+
+    final stmt = opened.prepare('select * from t');
+    expect(stmt.select(), [
+      {'c1': 1}
+    ]);
+
+    opened.execute('alter table t add column c2 default 2');
+
+    expect(stmt.select(), [
+      {'c1': 1, 'c2': 2}
+    ]);
+  });
+
   group('cursors', () {
     late CommonDatabase database;
 
@@ -287,6 +305,29 @@ void testPreparedStatements(
       expect(cursor.current, {'r': 2});
 
       expect(cursor.moveNext, throwsA(isA<SqliteException>()));
+      expect(cursor.moveNext(), isFalse);
+    });
+
+    test('handle recompilations while not running', () {
+      final opened = sqlite3.openInMemory()
+        ..execute('create table t (c1)')
+        ..execute('insert into t values (1)');
+      addTearDown(opened.dispose);
+
+      final stmt = opened.prepare('select * from t');
+      var cursor = stmt.selectCursor();
+
+      expect(cursor.moveNext(), isTrue);
+      expect(cursor.current, {'c1': 1});
+      expect(cursor.moveNext(), isFalse);
+
+      opened.execute('alter table t add column c2 default 2');
+      cursor = stmt.selectCursor();
+      expect(cursor.columnNames, ['c1']);
+
+      expect(cursor.moveNext(), isTrue);
+      expect(cursor.columnNames, ['c1', 'c2']);
+      expect(cursor.current, {'c1': 1, 'c2': 2});
       expect(cursor.moveNext(), isFalse);
     });